REF: Remove internals.base
